cleanup(core): extract task input resolution
Amp-Thread-ID: https://ampcode.com/threads/T-01a03515-99b0-703b-9319-59f48eea1bc0
